Fix bug that pmon can't load kernle from large disk partion
If the disk partion is larger than 120G, these errors  will occure:
"There is no linux partion" or "we can't locate root directory in super
block!". Now This bug is fixed. It should be usefull for all branches.

Thanks XuHaoyu <xuhaoyu@loongson.c>

+++ sys/sys/types.h
@@ -67,7 +67,7 @@ typedef	int64_t		quad_t;
 typedef	quad_t *	qaddr_t;
 
 typedef	char *		caddr_t;	/* core address */
-typedef	int32_t		daddr_t;	/* disk address */
+typedef	int64_t		daddr_t;	/* disk address */
 typedef	int32_t		dev_t;		/* device number */
 typedef	u_int32_t	fixpt_t;	/* fixed point number */
 typedef	u_int32_t	gid_t;		/* group id */
